sprint7-8/demo6/pyproject.toml
[tool.poetry]
name = "demo6"
version = "0.1.0"
description = ""
authors = ["Chrystian <chrystian@kenzie.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
